Add storage to the ty module.
